Fixes: Bundler app bundler error and infinite loop in activity_log
The app bundler error is fixed by moving the better-sqlite db driver,
which is node only, to a unit test that's blocklisted.

The infinite loop bug is fixed by using the correct sytance for
useFocusEffect and correctly mocking it in my unit test.

With these changes, I think I've finally gotten my implementation to
the point where it's working smoothly in tests and in the app.